Remove formatting warning in Mutation files
This commit fixes the formatting warning in the Mutation classes.

- For mut_type() and sense() calls, we changed the PRId8 format specification
  to PRId32
- For the -1 constants, we explicitly cast them to the right size
  (int8_t or int16_t) to reflect the format specification

+++ src/libaevol/mutation/Deletion.cpp
@@ -69,10 +69,10 @@ void Deletion::load(gzFile backup_file) {
 
 void Deletion::generic_description_string(char* str) const {
   // TODO <david.parsons@inria.fr> Use snprintf to avoid buffer overflow
-  sprintf(str, "%" PRId8 " %" PRId32 " %" PRId32 " %" PRId32 " %" PRId32
+  sprintf(str, "%" PRId32 " %" PRId32 " %" PRId32 " %" PRId32 " %" PRId32
       " %" PRId8 " %" PRId16 " %" PRId16 " %" PRId32 " %" PRId32,
-          mut_type(), pos1(), pos2(), -1, -1, -1,
-          align_score(), -1, length_, -1);
+          mut_type(), pos1(), pos2(), -1, -1, (int8_t) -1,
+          align_score(), (int16_t) -1, length_, -1);
   }
 
 // ============================================================================
